#530: Move ItemCategory endpoints to correct location
