+++ Api/Xipona.Api.Endpoints.Tests/v1/Controllers/ItemCategoryEndpointTests/DeleteItemCategoryTests.cs
@@ -5,7 +5,7 @@ using ProjectHermes.Xipona.Api.ApplicationServices.ItemCategories.Commands.Delet
 using ProjectHermes.Xipona.Api.Contracts.Common;
 using ProjectHermes.Xipona.Api.Domain.Common.Reasons;
 using ProjectHermes.Xipona.Api.Domain.TestKit.Common;
-using ProjectHermes.Xipona.Api.Endpoint.v1.Controllers;
+using ProjectHermes.Xipona.Api.Endpoint.v1.Endpoints;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common;
 using ProjectHermes.Xipona.Api.Endpoints.Tests.Common.StatusResults;
 using ProjectHermes.Xipona.Api.TestTools.Exceptions;
@@ -62,7 +62,7 @@ public class DeleteItemCategoryTests : EndpointCommandTestsBase<Guid, DeleteItem
         public override async Task<IResult> ExecuteTestMethod()
         {
             TestPropertyNotSetException.ThrowIfNull(_id);
-            return await MinimalItemCategoryController.DeleteItemCategory(
+            return await ItemCategoryEndpoints.DeleteItemCategory(
                 _id.Value,
                 CommandDispatcherMock.Object,
                 ErrorConverterMock.Object,
